renamed files (service folder) + added test mode to textGenerator
